# cargo-ship-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户注册成功"
          },
          "400": {
            "description": "输入数据验证失败"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        },
        "summary": "用户注册",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功，返回访问令牌和刷新令牌"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        },
        "summary": "用户登录",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功刷新访问令牌"
          },
          "401": {
            "description": "刷新令牌无效或已过期"
          }
        },
        "summary": "刷新访问令牌",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "登出成功"
          },
          "401": {
            "description": "未提供或无效的访问令牌"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "用户登出",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功"
          },
          "400": {
            "description": "旧密码错误或新密码不符合要求"
          },
          "401": {
            "description": "未提供或无效的访问令牌"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改当前用户密码",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "未提供或无效的访问令牌"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前用户信息",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/users": {
      "get": {
        "operationId": "AuthController_findAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有用户列表 (管理员)",
        "tags": [
          "认证与授权 (Auth)"
        ]
      },
      "post": {
        "operationId": "AuthController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "输入数据验证失败"
          },
          "403": {
            "description": "权限不足"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建新用户 (管理员)",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
    "/api/auth/users/{id}": {
      "get": {
        "operationId": "AuthController_findUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据ID获取单个用户 (管理员)",
        "tags": [
          "认证与授权 (Auth)"
        ]
      },
      "put": {
        "operationId": "AuthController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          },
          "409": {
            "description": "用户名或邮箱与他人冲突"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新用户信息 (管理员)",
        "tags": [
          "认证与授权 (Auth)"
        ]
      },
      "delete": {
        "operationId": "AuthController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户删除成功"
          },
          "400": {
            "description": "不能删除最后一个管理员"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除用户 (管理员)",
        "tags": [
          "认证与授权 (Auth)"
        ]
      }
    },
...
  },
  "info": {
    "title": "TDuCargoShipsManagement API",
    "description": "船舶货物管理系统 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名 (3-50位, 字母/数字/下划线/连字符)",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱地址",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "用户密码 (至少8位,包含大小写字母、数字和特殊字符)",
            "example": "Password123!"
          },
          "fullName": {
            "type": "string",
            "description": "用户全名",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "手机号码 (可选)",
            "example": "+86 138-1234-5678"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "fullName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名或邮箱地址",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "password123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "用于获取新访问令牌的刷新令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "当前使用的旧密码",
            "example": "OldPassword123!"
          },
          "newPassword": {
            "type": "string",
            "description": "符合强度要求的新密码",
            "example": "NewPassword123!"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {}
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户唯一标识符",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "john.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "用户全名",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "手机号码 (可选)",
            "example": "+86 138-1234-5678",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "active",
              "inactive",
              "locked",
              "pending"
            ],
            "example": "active"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "description": "最后登录时间",
            "nullable": true
          },
          "lastLoginIp": {
            "type": "string",
            "description": "最后登录IP地址",
            "example": "192.168.1.100",
            "nullable": true
          },
          "passwordChangedAt": {
            "format": "date-time",
            "type": "string",
            "description": "密码最后修改时间",
            "nullable": true
          },
          "roles": {
            "description": "用户拥有的角色列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "记录创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "记录最后更新时间"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "fullName",
          "status",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名 (3-50位, 字母/数字/下划线/连字符)",
            "example": "new_operator"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱地址",
            "example": "operator@example.com"
          },
          "password": {
            "type": "string",
            "description": "用户初始密码 (至少8位,包含大小写字母、数字和特殊字符)",
            "example": "Operator123!"
          },
          "fullName": {
            "type": "string",
            "description": "用户全名",
            "example": "李四"
          },
          "phoneNumber": {
            "type": "string",
            "description": "手机号码 (可选)",
            "example": "+86 139-8765-4321"
          },
          "roleIds": {
            "description": "分配给用户的角色ID数组",
            "example": [
              "role-uuid-operator"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "用户状态 (可选, 默认为 active)",
            "enum": [
              "active",
              "inactive",
              "locked",
              "pending"
            ],
            "example": "active"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "fullName",
          "roleIds"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "新用户名 (3-50位, 字母/数字/下划线/连字符)",
            "example": "john_doe_new"
          },
          "email": {
            "type": "string",
            "description": "新邮箱地址",
            "example": "john.doe.new@example.com"
          },
          "password": {
            "type": "string",
            "description": "重置后的新密码 (至少8位,包含大小写字母、数字和特殊字符)",
            "example": "NewPassword456!"
          },
          "fullName": {
            "type": "string",
            "description": "新用户全名",
            "example": "Johnathan Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "新手机号码",
            "example": "+1 555-123-4567"
          },
          "roleIds": {
            "description": "完全替换用户现有角色的ID数组",
            "example": [
              "role-uuid-viewer"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "新用户状态 (active, inactive, locked)",
            "enum": [
              "active",
              "inactive",
              "locked",
              "pending"
            ],
            "example": "inactive"
          }
        }
      },
...
    }
  }
}